refactor(ipc): split app payload schemas by domain
